requirements.txt
click==6.6
oauthlib==1.1.2
ordereddict==1.1
python-dotenv==0.5.1
requests==2.11.0
requests-oauthlib==0.6.2
six==1.10.0
tweepy==3.5.0
